Fix Linux linkage
