+++ lib/src/Utils/Utils.cpp
@@ -1,6 +1,7 @@
 // OpenCL Utils includes
 #include <CL/Utils/Utils.hpp>
 
+#include "Error.cpp"
 #include "Platform.cpp"
 #include "Device.cpp"
 #include "Context.cpp"
